# New weights using i7 machine