--- a/bin/node/executor/tests/submit_transaction.rs
+++ b/bin/node/executor/tests/submit_transaction.rs
@@ -230,7 +230,7 @@ fn submitted_transaction_should_be_valid() {
 		let res = Executive::validate_transaction(source, extrinsic);
 
 		assert_eq!(res.unwrap(), ValidTransaction {
-			priority: 1_411_390_000_000,
+			priority: 1_410_625_000_000,
 			requires: vec![],
 			provides: vec![(address, 0).encode()],
 			longevity: 128,