--- a/frame/utility/src/lib.rs
+++ b/frame/utility/src/lib.rs
@@ -205,17 +205,17 @@ mod weight_of {
 	use super::*;
 
 	/// - Base Weight:
-	///     - Create: 137.5 + 0.274 * S µs
-	///     - Approve: 103.8 + .266 * S µs
-	///     - Complete: 116.2 + .754 * S µs
+	///     - Create: 59.2 + 0.096 * S µs
+	///     - Approve: 42.27 + .116 * S µs
+	///     - Complete: 50.91 + .232 * S µs
 	/// - DB Weight:
 	///     - Reads: Multisig Storage, [Caller Account]
 	///     - Writes: Multisig Storage, [Caller Account]
 	/// - Plus Call Weight
 	pub fn as_multi<T: Trait>(other_sig_len: usize, call_weight: Weight) -> Weight {
 		call_weight
-			.saturating_add(150_000_000)
-			.saturating_add((other_sig_len as Weight).saturating_mul(750_000))
+			.saturating_add(60_000_000)
+			.saturating_add((other_sig_len as Weight).saturating_mul(250_000))
 			.saturating_add(T::DbWeight::get().reads_writes(1, 1))
 	}
 }
@@ -236,7 +236,7 @@ decl_module! {
 		/// - `calls`: The calls to be dispatched from the same origin.
 		///
 		/// # <weight>
-		/// - Base weight: 63.78 µs
+		/// - Base weight: 15.64 + .987 * c µs
 		/// - Plus the sum of the weights of the `calls`.
 		/// - Plus one additional event. (repeat read/write)
 		/// # </weight>
@@ -250,7 +250,7 @@ decl_module! {
 			|args: (&Vec<<T as Trait>::Call>,)| {
 				args.0.iter()
 					.map(|call| call.get_dispatch_info().weight)
-					.fold(65_000_000, |a: Weight, n| a.saturating_add(n))
+					.fold(15_000_000, |a: Weight, n| a.saturating_add(n).saturating_add(1_000_000))
 			},
 			|args: (&Vec<<T as Trait>::Call>,)| {
 				let all_operational = args.0.iter()
@@ -280,12 +280,12 @@ decl_module! {
 		/// The dispatch origin for this call must be _Signed_.
 		///
 		/// # <weight>
-		/// - Base weight: 5.1 µs
+		/// - Base weight: 2.863 µs
 		/// - Plus the weight of the `call`
 		/// # </weight>
 		#[weight = FunctionOf(
 			|args: (&u16, &Box<<T as Trait>::Call>)| {
-				args.1.get_dispatch_info().weight.saturating_add(5_000_000)
+				args.1.get_dispatch_info().weight.saturating_add(3_000_000)
 			},
 			|args: (&u16, &Box<<T as Trait>::Call>)| args.1.get_dispatch_info().class,
 			Pays::Yes,
@@ -339,9 +339,9 @@ decl_module! {
 		///   `MultisigDepositBase + threshold * MultisigDepositFactor`.
 		/// -------------------------------
 		/// - Base Weight:
-		///     - Create: 137.5 + 0.274 * S µs
-		///     - Approve: 103.8 + .266 * S µs
-		///     - Complete: 116.2 + .754 * S µs
+		///     - Create: 59.2 + 0.096 * S µs
+		///     - Approve: 42.27 + .116 * S µs
+		///     - Complete: 50.91 + .232 * S µs
 		/// - DB Weight:
 		///     - Reads: Multisig Storage, [Caller Account]
 		///     - Writes: Multisig Storage, [Caller Account]
@@ -471,8 +471,8 @@ decl_module! {
 		///   `MultisigDepositBase + threshold * MultisigDepositFactor`.
 		/// ----------------------------------
 		/// - Base Weight:
-		///     - Create: 139.1 + 0.202 * S
-		///     - Approve: 96.6 + 0.328 * S
+		///     - Create: 56.3 + 0.107 * S
+		///     - Approve: 39.25 + 0.121 * S
 		/// - DB Weight:
 		///     - Read: Multisig Storage, [Caller Account]
 		///     - Write: Multisig Storage, [Caller Account]
@@ -480,8 +480,8 @@ decl_module! {
 		#[weight = FunctionOf(
 			|args: (&u16, &Vec<T::AccountId>, &Option<Timepoint<T::BlockNumber>>, &[u8; 32])| {
 				T::DbWeight::get().reads_writes(1, 1)
-					.saturating_add(140_000_000)
-					.saturating_add((args.1.len() as Weight).saturating_mul(350_000))
+					.saturating_add(60_000_000)
+					.saturating_add((args.1.len() as Weight).saturating_mul(120_000))
 			},
 			DispatchClass::Normal,
 			Pays::Yes,
@@ -554,7 +554,7 @@ decl_module! {
 		/// - I/O: 1 read `O(S)`, one remove.
 		/// - Storage: removes one item.
 		/// ----------------------------------
-		/// - Base Weight: 126.6 + 0.126 * S
+		/// - Base Weight: 46.71 + 0.09 * S
 		/// - DB Weight:
 		///     - Read: Multisig Storage, [Caller Account]
 		///     - Write: Multisig Storage, [Caller Account]
@@ -562,8 +562,8 @@ decl_module! {
 		#[weight = FunctionOf(
 			|args: (&u16, &Vec<T::AccountId>, &Timepoint<T::BlockNumber>, &[u8; 32])| {
 				T::DbWeight::get().reads_writes(1, 1)
-					.saturating_add(130_000_000)
-					.saturating_add((args.1.len() as Weight).saturating_mul(130_000))
+					.saturating_add(50_000_000)
+					.saturating_add((args.1.len() as Weight).saturating_mul(100_000))
 			},
 			DispatchClass::Normal,
 			Pays::Yes,