--- a/frame/vesting/src/lib.rs
+++ b/frame/vesting/src/lib.rs
@@ -193,10 +193,12 @@ decl_module! {
 		/// - DbWeight: 2 Reads, 2 Writes
 		///     - Reads: Vesting Storage, Balances Locks, [Sender Account]
 		///     - Writes: Vesting Storage, Balances Locks, [Sender Account]
-		/// - Benchmark: 147.5 µs (min square analysis)
-		/// - Assuming less than 50 locks on any user, else we may want factor in number of locks.
+		/// - Benchmark:
+		///     - Unlocked: 56.1 + .098 * l µs (min square analysis)
+		///     - Locked: 54.37 + .254 * l µs (min square analysis)
+		/// - Using 60 µs fixed. Assuming less than 50 locks on any user, else we may want factor in number of locks.
 		/// # </weight>
-		#[weight = 150_000_000 + T::DbWeight::get().reads_writes(2, 2)]
+		#[weight = 60_000_000 + T::DbWeight::get().reads_writes(2, 2)]
 		fn vest(origin) -> DispatchResult {
 			let who = ensure_signed(origin)?;
 			Self::update_lock(who)
@@ -216,10 +218,12 @@ decl_module! {
 		/// - DbWeight: 3 Reads, 3 Writes
 		///     - Reads: Vesting Storage, Balances Locks, Target Account
 		///     - Writes: Vesting Storage, Balances Locks, Target Account
-		/// - Benchmark: 150.4 µs (min square analysis)
-		/// - Assuming less than 50 locks on any user, else we may want factor in number of locks.
+		/// - Benchmark:
+		///     - Unlocked: 58.09 + .104 * l µs (min square analysis)
+		///     - Locked: 55.35 + .255 * l µs (min square analysis)
+		/// - Using 60 µs fixed. Assuming less than 50 locks on any user, else we may want factor in number of locks.
 		/// # </weight>
-		#[weight = 150_000_000 + T::DbWeight::get().reads_writes(3, 3)]
+		#[weight = 60_000_000 + T::DbWeight::get().reads_writes(3, 3)]
 		fn vest_other(origin, target: <T::Lookup as StaticLookup>::Source) -> DispatchResult {
 			ensure_signed(origin)?;
 			Self::update_lock(T::Lookup::lookup(target)?)
@@ -240,10 +244,10 @@ decl_module! {
 		/// - DbWeight: 3 Reads, 3 Writes
 		///     - Reads: Vesting Storage, Balances Locks, Target Account, [Sender Account]
 		///     - Writes: Vesting Storage, Balances Locks, Target Account, [Sender Account]
-		/// - Benchmark: 263 µs (min square analysis)
-		/// - Assuming less than 50 locks on any user, else we may want factor in number of locks.
+		/// - Benchmark: 111.4 + .345 * l µs (min square analysis)
+		/// - Using 115 µs fixed. Assuming less than 50 locks on any user, else we may want factor in number of locks.
 		/// # </weight>
-		#[weight = 300_000_000 + T::DbWeight::get().reads_writes(3, 3)]
+		#[weight = 115_000_000 + T::DbWeight::get().reads_writes(3, 3)]
 		pub fn vested_transfer(
 			origin,
 			target: <T::Lookup as StaticLookup>::Source,