We are using a faster machine than before to re-calculate our weights, which results in a drop in weight values of around 2x - 3x.

This PR updates weights for:

* DbWeight
* Utility Pallet
* Vesting Pallet
* Balances Pallet

Current machine specifications:

<details>
  <summary><code>lscpu - Intel(R) Core(TM) i7-7700K CPU @ 4.20GHz (expand)</code></summary>

```
shawntabrizi@bm2:~$ lscpu
Architecture:                    x86_64
CPU op-mode(s):                  32-bit, 64-bit
Byte Order:                      Little Endian
Address sizes:                   39 bits physical, 48 bits virtual
CPU(s):                          8
On-line CPU(s) list:             0-7
Thread(s) per core:              2
Core(s) per socket:              4
Socket(s):                       1
NUMA node(s):                    1
Vendor ID:                       GenuineIntel
CPU family:                      6
Model:                           158
Model name:                      Intel(R) Core(TM) i7-7700K CPU @ 4.20GHz
Stepping:                        9
CPU MHz:                         801.237
CPU max MHz:                     4500.0000
CPU min MHz:                     800.0000
BogoMIPS:                        8400.00
Virtualization:                  VT-x
L1d cache:                       128 KiB
L1i cache:                       128 KiB
L2 cache:                        1 MiB
L3 cache:                        8 MiB
NUMA node0 CPU(s):               0-7
Vulnerability Itlb multihit:     KVM: Mitigation: Split huge pages
Vulnerability L1tf:              Mitigation; PTE Inversion; VMX conditional cache flushes, SMT vulnerable
Vulnerability Mds:               Mitigation; Clear CPU buffers; SMT vulnerable
Vulnerability Meltdown:          Mitigation; PTI
Vulnerability Spec store bypass: Mitigation; Speculative Store Bypass disabled via prctl and seccomp
Vulnerability Spectre v1:        Mitigation; usercopy/swapgs barriers and __user pointer sanitization
Vulnerability Spectre v2:        Mitigation; Full generic retpoline, IBPB conditional, IBRS_FW, STIBP conditional, RSB filling
Vulnerability Tsx async abort:   Mitigation; Clear CPU buffers; SMT vulnerable
Flags:                           fpu vme de pse tsc msr pae mce cx8 apic sep mtrr pge mca cmov pat pse36 clflush dts acpi mmx fxsr sse sse2 ss ht tm pbe syscall nx pdpe1gb rdtscp lm constant_tsc 
                                 art arch_perfmon pebs bts rep_good nopl xtopology nonstop_tsc cpuid aperfmperf pni pclmulqdq dtes64 monitor ds_cpl vmx est tm2 ssse3 sdbg fma cx16 xtpr pdcm pcid 
                                 sse4_1 sse4_2 x2apic movbe popcnt tsc_deadline_timer aes xsave avx f16c rdrand lahf_lm abm 3dnowprefetch cpuid_fault epb invpcid_single pti ssbd ibrs ibpb stibp t
                                 pr_shadow vnmi flexpriority ept vpid ept_ad fsgsbase tsc_adjust bmi1 hle avx2 smep bmi2 erms invpcid rtm mpx rdseed adx smap clflushopt intel_pt xsaveopt xsavec x
                                 getbv1 xsaves dtherm ida arat pln pts hwp hwp_notify hwp_act_window hwp_epp md_clear flush_l1d
```
</details>


<details>
  <summary><code>lshw (expand)</code></summary>

```
shawntabrizi@bm2:~$ sudo lshw
bm2                         
    description: Rack Mount Chassis
    product: S1200SP (SKU Number)
    vendor: Intel Corporation
    version: ....................
    serial: ............
    width: 64 bits
    capabilities: smbios-2.7 dmi-2.7 smp vsyscall32
    configuration: administrator_password=disabled boot=normal chassis=rackmount family=Family frontpanel_password=disabled sku=SKU Number uuid=957C3907-F8C9-11E6-AAD1-A4BF01182C4E
  *-core
       description: Motherboard
       product: S1200SP
       vendor: Intel Corporation
       physical id: 0
       version: H57532-250
       serial: QSSA70704838
       slot: Part Component
     *-firmware
          description: BIOS
          vendor: Intel Corporation
          physical id: 6
          version: S1200SP.86B.03.01.0042.013020190050
          date: 01/30/2019
          size: 64KiB
          capacity: 16MiB
          capabilities: pci pnp upgrade shadowing cdboot bootselect edd int13floppy1200 int13floppy720 int13floppy2880 int5printscreen int9keyboard int14serial int17printer int10video acpi usb ls120boot zipboot biosbootspecification netboot uefi
     *-cpu
          description: CPU
          product: Intel(R) Core(TM) i7-7700K CPU @ 4.20GHz
          vendor: Intel Corp.
          physical id: 1d
          bus info: cpu@0
          version: Intel(R) Core(TM) i7-7700K CPU @ 4.20GHz
          serial: To Be Filled By O.E.M.
          slot: CPU 1
          size: 4388MHz
          capacity: 4500MHz
          width: 64 bits
          clock: 100MHz
          capabilities: lm fpu fpu_exception wp vme de pse tsc msr pae mce cx8 apic sep mtrr pge mca cmov pat pse36 clflush dts acpi mmx fxsr sse sse2 ss ht tm pbe syscall nx pdpe1gb rdtscp x86-64 constant_tsc art arch_perfmon pebs bts rep_good nopl xtopology nonstop_tsc cpuid aperfmperf pni pclmulqdq dtes64 monitor ds_cpl vmx est tm2 ssse3 sdbg fma cx16 xtpr pdcm pcid sse4_1 sse4_2 x2apic movbe popcnt tsc_deadline_timer aes xsave avx f16c rdrand lahf_lm abm 3dnowprefetch cpuid_fault epb invpcid_single pti ssbd ibrs ibpb stibp tpr_shadow vnmi flexpriority ept vpid ept_ad fsgsbase tsc_adjust bmi1 hle avx2 smep bmi2 erms invpcid rtm mpx rdseed adx smap clflushopt intel_pt xsaveopt xsavec xgetbv1 xsaves dtherm ida arat pln pts hwp hwp_notify hwp_act_window hwp_epp md_clear flush_l1d cpufreq
          configuration: cores=4 enabledcores=4 threads=8
        *-cache:0
             description: L1 cache
             physical id: 1a
             slot: L1 Cache
             size: 128KiB
             capacity: 128KiB
             capabilities: synchronous internal write-through instruction
             configuration: level=1
        *-cache:1
             description: L2 cache
             physical id: 1b
             slot: L2 Cache
             size: 1MiB
             capacity: 1MiB
             capabilities: synchronous internal write-through unified
             configuration: level=2
        *-cache:2
             description: L3 cache
             physical id: 1c
             slot: L3 Cache
             size: 8MiB
             capacity: 8MiB
             capabilities: synchronous internal write-back unified
             configuration: level=3
     *-cache
          description: L1 cache
          physical id: 19
          slot: L1 Cache
          size: 128KiB
          capacity: 128KiB
          capabilities: synchronous internal write-through data
          configuration: level=1
     *-memory
          description: System Memory
          physical id: 1e
          slot: System board or motherboard
          size: 64GiB
        *-bank:0
             description: DIMM DDR4 Synchronous 2400 MHz (0.4 ns)
             product: KHX2400C15/16G
             vendor: Kingston
             physical id: 0
             serial: E81EC776
             slot: DIMM_A1
             size: 16GiB
             width: 64 bits
             clock: 2400MHz (0.4ns)
        *-bank:1
             description: DIMM DDR4 Synchronous 2400 MHz (0.4 ns)
             product: KHX2400C15/16G
             vendor: Kingston
             physical id: 1
             serial: 0071E8783
             slot: DIMM_A2
             size: 16GiB
             width: 64 bits
             clock: 2400MHz (0.4ns)
        *-bank:2
             description: DIMM DDR4 Synchronous 2400 MHz (0.4 ns)
             product: KHX2400C15/16G
             vendor: Kingston
             physical id: 2
             serial: ED1EC676
             slot: DIMM_B1
             size: 16GiB
             width: 64 bits
             clock: 2400MHz (0.4ns)
        *-bank:3
             description: DIMM DDR4 Synchronous 2400 MHz (0.4 ns)
             product: KHX2400C15/16G
             vendor: Kingston
             physical id: 3
             serial: 00C1EBB83
             slot: DIMM_B2
             size: 16GiB
             width: 64 bits
             clock: 2400MHz (0.4ns)
     *-pci
          description: Host bridge
          product: Xeon E3-1200 v6/7th Gen Core Processor Host Bridge/DRAM Registers
          vendor: Intel Corporation
          physical id: 100
          bus info: pci@0000:00:00.0
          version: 05
          width: 32 bits
          clock: 33MHz
          configuration: driver=skl_uncore
          resources: irq:0
        *-pci:0
             description: PCI bridge
             product: Xeon E3-1200 v5/E3-1500 v5/6th Gen Core Processor PCIe Controller (x16)
             vendor: Intel Corporation
             physical id: 1
             bus info: pci@0000:00:01.0
             version: 05
             width: 32 bits
             clock: 33MHz
             capabilities: pci pm msi pciexpress normal_decode bus_master cap_list
             configuration: driver=pcieport
             resources: irq:122
        *-pci:1
             description: PCI bridge
             product: Xeon E3-1200 v5/E3-1500 v5/6th Gen Core Processor PCIe Controller (x8)
             vendor: Intel Corporation
             physical id: 1.1
             bus info: pci@0000:00:01.1
             version: 05
             width: 32 bits
             clock: 33MHz
             capabilities: pci pm msi pciexpress normal_decode bus_master cap_list
             configuration: driver=pcieport
             resources: irq:123 memory:a2e00000-a2efffff
           *-storage
                description: Non-Volatile memory controller
                product: PCIe Data Center SSD
                vendor: Intel Corporation
                physical id: 0
                bus info: pci@0000:02:00.0
                version: 02
                width: 64 bits
                clock: 33MHz
                capabilities: storage pm msix pciexpress nvm_express bus_master cap_list rom
                configuration: driver=nvme latency=0
                resources: irq:17 memory:a2e00000-a2e03fff memory:a2e10000-a2e1ffff
        *-display UNCLAIMED
             description: Display controller
             product: HD Graphics 630
             vendor: Intel Corporation
             physical id: 2
             bus info: pci@0000:00:02.0
             version: 04
             width: 64 bits
             clock: 33MHz
             capabilities: pciexpress msi pm bus_master cap_list
             configuration: latency=0
             resources: memory:a1000000-a1ffffff memory:90000000-9fffffff ioport:5000(size=64)
        *-usb
             description: USB controller
             product: 100 Series/C230 Series Chipset Family USB 3.0 xHCI Controller
             vendor: Intel Corporation
             physical id: 14
             bus info: pci@0000:00:14.0
             version: 31
             width: 64 bits
             clock: 33MHz
             capabilities: pm msi xhci bus_master cap_list
             configuration: driver=xhci_hcd latency=0
             resources: irq:128 memory:a2f00000-a2f0ffff
           *-usbhost:0
                product: xHCI Host Controller
                vendor: Linux 5.4.0-26-generic xhci-hcd
                physical id: 0
                bus info: usb@1
                logical name: usb1
                version: 5.04
                capabilities: usb-2.00
                configuration: driver=hub slots=16 speed=480Mbit/s
              *-usb
                   description: Keyboard
                   product: Virtual Keyboard and Mouse
                   vendor: American Megatrends Inc.
                   physical id: 1
                   bus info: usb@1:1
                   version: 1.00
                   serial: serial
                   capabilities: usb-2.00
                   configuration: driver=usbhid speed=12Mbit/s
           *-usbhost:1
                product: xHCI Host Controller
                vendor: Linux 5.4.0-26-generic xhci-hcd
                physical id: 1
                bus info: usb@2
                logical name: usb2
                version: 5.04
                capabilities: usb-3.00
                configuration: driver=hub slots=10 speed=5000Mbit/s
        *-generic UNCLAIMED
             description: Signal processing controller
             product: 100 Series/C230 Series Chipset Family Thermal Subsystem
             vendor: Intel Corporation
             physical id: 14.2
             bus info: pci@0000:00:14.2
             version: 31
             width: 64 bits
             clock: 33MHz
             capabilities: pm msi bus_master cap_list
             configuration: latency=0
             resources: iomemory:100-ff memory:10bff01000-10bff01fff
        *-communication:0 UNCLAIMED
             description: Communication controller
             product: 100 Series/C230 Series Chipset Family MEI Controller #1
             vendor: Intel Corporation
             physical id: 16
             bus info: pci@0000:00:16.0
             version: 31
             width: 64 bits
             clock: 33MHz
             capabilities: pm msi cap_list
             configuration: latency=0
             resources: memory:a2f17000-a2f17fff
        *-communication:1 UNCLAIMED
             description: Communication controller
             product: 100 Series/C230 Series Chipset Family MEI Controller #2
             vendor: Intel Corporation
             physical id: 16.1
             bus info: pci@0000:00:16.1
             version: 31
             width: 64 bits
             clock: 33MHz
             capabilities: pm msi bus_master cap_list
             configuration: latency=0
             resources: memory:a2f16000-a2f16fff
        *-sata
             description: SATA controller
             product: Q170/Q150/B150/H170/H110/Z170/CM236 Chipset SATA Controller [AHCI Mode]
             vendor: Intel Corporation
             physical id: 17
             bus info: pci@0000:00:17.0
             version: 31
             width: 32 bits
             clock: 66MHz
             capabilities: sata msi pm ahci_1.0 bus_master cap_list
             configuration: driver=ahci latency=0
             resources: irq:129 memory:a2f14000-a2f15fff memory:a2f19000-a2f190ff ioport:5080(size=8) ioport:5088(size=4) ioport:5060(size=32) memory:a2f18000-a2f187ff
        *-pci:2
             description: PCI bridge
             product: 100 Series/C230 Series Chipset Family PCI Express Root Port #5
             vendor: Intel Corporation
             physical id: 1c
             bus info: pci@0000:00:1c.0
             version: f1
             width: 32 bits
             clock: 33MHz
             capabilities: pci pciexpress msi pm normal_decode bus_master cap_list
             configuration: driver=pcieport
             resources: irq:124 memory:a2d00000-a2dfffff
           *-storage
                description: Non-Volatile memory controller
                product: PCIe Data Center SSD
                vendor: Intel Corporation
                physical id: 0
                bus info: pci@0000:03:00.0
                version: 02
                width: 64 bits
                clock: 33MHz
                capabilities: storage pm msix pciexpress nvm_express bus_master cap_list rom
                configuration: driver=nvme latency=0
                resources: irq:16 memory:a2d00000-a2d03fff memory:a2d10000-a2d1ffff
        *-pci:3
             description: PCI bridge
             product: 100 Series/C230 Series Chipset Family PCI Express Root Port #10
             vendor: Intel Corporation
             physical id: 1d
             bus info: pci@0000:00:1d.0
             version: f1
             width: 32 bits
             clock: 33MHz
             capabilities: pci pciexpress msi pm normal_decode bus_master cap_list
             configuration: driver=pcieport
             resources: irq:125 memory:a2000000-a28fffff ioport:a0000000(size=16777216)
           *-display UNCLAIMED
                description: VGA compatible controller
                product: MGA G200e [Pilot] ServerEngines (SEP1)
                vendor: Matrox Electronics Systems Ltd.
                physical id: 0
                bus info: pci@0000:04:00.0
                version: 05
                width: 32 bits
                clock: 33MHz
                capabilities: pm pciexpress msi vga_controller bus_master cap_list
                configuration: latency=0
                resources: memory:a0000000-a0ffffff memory:a2800000-a2803fff memory:a2000000-a27fffff memory:a2810000-a281ffff
        *-pci:4
             description: PCI bridge
             product: 100 Series/C230 Series Chipset Family PCI Express Root Port #11
             vendor: Intel Corporation
             physical id: 1d.2
             bus info: pci@0000:00:1d.2
             version: f1
             width: 32 bits
             clock: 33MHz
             capabilities: pci pciexpress msi pm normal_decode bus_master cap_list
             configuration: driver=pcieport
             resources: irq:126 ioport:4000(size=4096) memory:a2b00000-a2cfffff
           *-network
                description: Ethernet interface
                product: I210 Gigabit Network Connection
                vendor: Intel Corporation
                physical id: 0
                bus info: pci@0000:05:00.0
                logical name: eno1
                version: 03
                serial: a4:bf:01:18:2c:4e
                size: 1Gbit/s
                capacity: 1Gbit/s
                width: 32 bits
                clock: 33MHz
                capabilities: pm msi msix pciexpress bus_master cap_list ethernet physical tp 10bt 10bt-fd 100bt 100bt-fd 1000bt-fd autonegotiation
                configuration: autonegotiation=on broadcast=yes driver=igb driverversion=5.6.0-k duplex=full firmware=3.25, 0x8000065c ip=149.202.69.202 latency=0 link=yes multicast=yes port=twisted pair speed=1Gbit/s
                resources: irq:18 memory:a2b00000-a2bfffff ioport:4000(size=32) memory:a2c00000-a2c03fff
        *-pci:5
             description: PCI bridge
             product: 100 Series/C230 Series Chipset Family PCI Express Root Port #12
             vendor: Intel Corporation
             physical id: 1d.3
             bus info: pci@0000:00:1d.3
             version: f1
             width: 32 bits
             clock: 33MHz
             capabilities: pci pciexpress msi pm normal_decode bus_master cap_list
             configuration: driver=pcieport
             resources: irq:127 ioport:3000(size=4096) memory:a2900000-a2afffff
           *-network DISABLED
                description: Ethernet interface
                product: I210 Gigabit Network Connection
                vendor: Intel Corporation
                physical id: 0
                bus info: pci@0000:06:00.0
                logical name: eno2
                version: 03
                serial: a4:bf:01:18:2c:4f
                capacity: 1Gbit/s
                width: 32 bits
                clock: 33MHz
                capabilities: pm msi msix pciexpress bus_master cap_list ethernet physical tp 10bt 10bt-fd 100bt 100bt-fd 1000bt-fd autonegotiation
                configuration: autonegotiation=on broadcast=yes driver=igb driverversion=5.6.0-k firmware=3.25, 0x8000065d latency=0 link=no multicast=yes port=twisted pair
                resources: irq:19 memory:a2900000-a29fffff ioport:3000(size=32) memory:a2a00000-a2a03fff
        *-isa
             description: ISA bridge
             product: C236 Chipset LPC/eSPI Controller
             vendor: Intel Corporation
             physical id: 1f
             bus info: pci@0000:00:1f.0
…