@shawntabrizi 2020-04-30
TODO follow up from this PR:

* Migrate other pallets to new weights (original owners will take care of that in seperate PRS)
* Update Extrinsic Base Weight
* Add Block Base Weight