@xlc 2020-04-30
So change benchmarking baseline machine means modify all the numbers?
Shouldn't be just modify the baseline number instead? i.e. previously 1 picosecond = 1 weight, now 1 picosecond = 2 weight?

Good that we just start working on writing benchmarking and not yet reach to running benchmarking phase. Otherwise we will be wasting our time.