@shawntabrizi 2020-04-30
@xlc we want to, at least to start, have 1_000 weight == 1 nanosecond.

In general, until we launch polkadot, you should not expect that anything stays fixed.

As Gav has mentioned, it is likely that the hardware specifications of weight calculation will change per major release, but for this first release, we are still exploring where we want to set the base level.