@xlc 2020-05-01
So this means we shouldn't start benchmarking yet as the result may not be consistent with Substrate?