@shawntabrizi 2020-05-01
@xlc You can def start to write benchmarking code, but running the final benchmarks and applying final weight values is something I might wait to do.