@gavofyork 2020-05-01
(Wait until 2.0 release, where we will have 100% fixed the standard hardware configuration. That said, this is almost certainly that configuration.)